worker/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
litellm==1.48.0
supabase==2.9.1
python-dotenv==1.0.1
cryptography==43.0.1
httpx==0.27.2
inngest==0.4.0
